Main.HeatTransfer.BC — Method

BC(func)

Specify the boundary condition with a temperature function of time func(t).

BC(u::Real)

Specify the boundary condition as a constant temperature u.

Main.HeatTransfer.solve_heat — Method

solve_heat(initial::AbstractVector; len, t, bc, a=1, every=1)::Matrix

Solve the heat equation for an initial temperature state. Returns a matrix with temperature states at each timestep in each column.

Arguments

initial::AbstractVector: Vector of initial temperatures.

len: The length of the 1-D material.

t::AbstractRange: Range of times to solve for, including the initial state.

bc::NTuple{2, BC}: Boundary conditions at the edges of the 1-D material.

a=1: Thermal diffusivity of the material.

every=1: Return one state for every every states in t. For example, every=2 saves every other state that is calculated.

Main.HeatTransfer.solve_heat — Method

solve_heat(initial, x::AbstractRange; t, bc, a=1, every=1)::Matrix

initial can be provided as a function of position initial(s) with x as a range of positions to use for calculations.

Main.HeatTransfer.solve_heat! — Method

solve_heat!(states::AbstractMatrix, initial::AbstractVector, len, t, bc, a=1, every=1)

solve_heat but store results in-place to states.

Missing docstring.

Missing docstring for L_times!(out::AbstractVector{T}, x::AbstractVector{T}) where T. Check Documenter's build log for details.